+++ core/src/main/java/com/asml/apa/wta/core/config/RuntimeConfig.java
@@ -1,6 +1,8 @@
 package com.asml.apa.wta.core.config;
 
 import com.asml.apa.wta.core.model.enums.Domain;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.AllArgsConstructor;
@@ -25,4 +27,17 @@ public class RuntimeConfig {
 
   @Builder.Default
   private Map<String, String> events = new HashMap<>();
+
+  @Builder.Default
+  private String logLevel = "INFO";
+
+  @Builder.Default
+  private boolean doConsoleLog = true;
+
+  @Builder.Default
+  private boolean doFileLog = true;
+
+  @Builder.Default
+  private String logPath =
+      "logs/" + LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd_HHmmss")) + ".log";
 }