+++ core/src/test/java/com/asml/apa/wta/core/ConfigReaderIntegrationTest.java
@@ -9,7 +9,7 @@ import com.asml.apa.wta.core.utils.WtaUtils;
 import java.util.HashMap;
 import org.junit.jupiter.api.Test;
 
-public class ConfigReaderTest {
+class ConfigReaderIntegrationTest {
   @Test
   void readsConfigFileCorrectly() {
     RuntimeConfig cr = WtaUtils.readConfig("src/test/resources/testConfig.json");
@@ -22,6 +22,9 @@ public class ConfigReaderTest {
         put("f2", "v2");
       }
     });
+    assertThat(cr.getLogLevel()).isEqualTo("INFO");
+    assertThat(cr.isDoConsoleLog()).isFalse();
+    assertThat(cr.isDoFileLog()).isTrue();
   }
 
   @Test
@@ -36,6 +39,9 @@ public class ConfigReaderTest {
         put("f2", "v2");
       }
     });
+    assertThat(cr.getLogLevel()).isEqualTo("INFO");
+    assertThat(cr.isDoConsoleLog()).isFalse();
+    assertThat(cr.isDoFileLog()).isTrue();
   }
 
   @Test
@@ -53,5 +59,22 @@ public class ConfigReaderTest {
           WtaUtils.readConfig("src/test/resources/testConfigInvalid.json");
         })
         .isInstanceOf(IllegalArgumentException.class);
+  }
+
+  @Test
+  void readsConfigFileWhereLogSettingIsNotThere() {
+    RuntimeConfig cr = WtaUtils.readConfig("src/test/resources/testConfigNoLogSettings.json");
+    assertThat(cr.getAuthors()).isEqualTo("Test Name");
+    assertThat(cr.getDomain()).isEqualTo("Test Domain");
+    assertThat(cr.getDescription()).isEqualTo("Test Description");
+    assertThat(cr.getEvents()).isEqualTo(new HashMap<String, String>() {
+      {
+        put("f1", "v1");
+        put("f2", "v2");
+      }
+    });
+    assertThat(cr.getLogLevel()).isEqualTo("INFO");
+    assertThat(cr.isDoConsoleLog()).isTrue();
+    assertThat(cr.isDoFileLog()).isTrue();
   }
 }